Configures console application outside listener
