+++ config/module.config.php
@@ -22,10 +22,10 @@ use Doctrine\DBAL\Tools\Console;
 use Doctrine\ORM\Tools\Console\Command;
 use DoctrineModule\Form\Element;
 use DoctrineModule\Stdlib\Hydrator\DoctrineObject;
+use DoctrineORMModule\CliConfigurator;
 use DoctrineORMModule\Listener\PostCliLoadListener;
 use DoctrineORMModule\Service;
 use DoctrineORMModule\Yuml;
-use Zend\ServiceManager\Factory\InvokableFactory;
 
 return [
     'doctrine' => [
@@ -197,7 +197,8 @@ return [
 
     'service_manager' => [
         'factories' => [
-            PostCliLoadListener::class => InvokableFactory::class,
+            PostCliLoadListener::class => Service\PostCliLoadListenerFactory::class,
+            CliConfigurator::class => Service\CliConfiguratorFactory::class,
             'Doctrine\ORM\EntityManager' => Service\EntityManagerAliasCompatFactory::class,
         ],
         'invokables' => [
